ingestion/requirements.txt
httpx==0.27.0
supabase==2.10.0
python-dotenv==1.0.1
tenacity==8.3.0
